Cancellation support for workflows (#60)
Implemented cancellation support for workflows and their parts.
The basic idea is that workflow code is always executed in a cancellation scope.
By default cancellation scopes are hierarchical. So any scope created explicitly is a child of a current one. A cancellation request to a scope automatically propagated to all its children.
To support cleanup in case of a cancellation a scope can be created in detached mode. Detached scope doesn't have a parent and doesn't receive implicit cancellation requests. It still can be cancelled explicitly.
Workflow thread is by definition is its own scope. So cancelling a thread automatically cancels all its non detached children.
A cancellation request is treated differently than the standard Java thread interrupts. Interrupts affect blocking operations. For example Future.get throws interrupted exception if thread is interrupted. It doesn't cause the operation that future waits on to get aborted. Cancellation affects operations directly. So for example a thread is waiting on a future.get that is result of activity execution. When thread is cancelled the "get" call is not interrupted. Activity execution is cancelled instead and after successful cancellation completes the future with CancellationException. Note that for this to work the operation itself should be in the cancelled scope, not the future.get(). The following operations that are currently obeying cancellation requests: activity executions, child workflows, timer, sleep.