Refactoring of exception hierarchies and propagation (#74)
The basic idea is that any exception thrown from an activity or child workflow is delivered as is to a calling workflow and wrapped in appropriate exception. So if NPE thrown in an activity executed from a child workflow the external client receives the following chain:

 WorkflowFailureException
     ->ChildWorkflowFailureException
             ->ActivityFailureException
                    ->NullPointerException
Also all exceptions thrown from an activity inherit from ActivityException and from ChildWorkflow inherit from ChildWorkflowException. Both of them inherit from WorkflowOperationException which is base for all exceptions that library throws. The only "exception" from this rule is CancellationException which is standard java java.util.concurrent one. And it is thrown from any operation that was cancelled.