Clean worker shutdown (#237)
Refactoring of the worker shutdown API and logic.
After the refactoring the worker shutdown is similar to the shutdown of a ThreadPoolExecutor. It is done through shutdown/shutdownNow and then awaitTermination. An activity heartbeat after shutdown/shutdownNow starts failing with ActivityWorkerShutdownException. shutdownNow also sends an interrupt to all currently executing activities..

Bumped the library version to 2.2.0