Feature/gradle 4 2 to 6 0 openjdk 8 alpine docker base image (#426)
* Upgrade gradle wrapper from 4.2 to 6.0 (#417)

To improve compilation time and fix numerous amount of issues alongside
other improvements that have been done throughout Gradle 5.0 and 6.0

* Update Gradle wrapper jar and properties (#417)
In order to work with the newest Gradle version, the Wrapper jar and
properties also have to be updated, so that Gradle knows how to download
it's executable
Gradle automatically downloads the Gradle executable through the
properties

* Introduce OpenJDK Alpine Docker base image (#417)
In order to have the smallest image possible, the Alpine image has to
be used. This will increase performance and will use less disk space
and network ingress.
This base image will reduce the final image size from 713MB down to
377MB, this all by using as minimum libraries as needed; for example by
removing the linux-gnu library.

* Update multiple Gradle dependencies+plugins (#417)
In order to keep up with the latest technology, benefit from the
fastest features and to improve security due to fixed issues and
bugs, there has to be a regular update for dependency and plugins.
Update multiple Gradle dependencies and plugins:
Plugins:
- thrift-gradle-plugin (0.4.0 -> 0.4.1)
- net.minecrell.licenser (0.3 -> 0.4.1)
- com.github.sherter.google-java-format (0.6 -> 0.8)
- net.ltgt.errorprone (0.6 -> 1.1.1)
- java -> java-library

Compile Dependencies:
com.google.errorprone:error_prone_core (2.3.1 -> 2.3.3)
com.uber.m3 (0.2.3 -> 0.4.0)
com.google.guava (27.0.1-jre -> 28.1-jre)

* Updated GSON dependency to 2.8.6 (uber#417)
In order to keep up with the latest technology, benefit from the
fastest features and to improve security due to fixed issues and
bugs, there has to be a regular update for dependency and plugins.
GSON 2.8.6 deprecates certain methods which will prevent usage of
these methods in the future.
https://github.com/google/gson/blob/master/CHANGELOG.md#version-286

* Removed deprecated JsonParser#parse (uber#417)
JsonParser#parse and the initialization of the JsonParser objects
were deprecated in version 2.8.6 of GSON.
In order to be complient with the newest version, the code was updated
to use the static JsonParser#parseString method.

* Updated cron-utils from 8.0.0 -> 9.0.0 (uber#417)
It has been over a year since the version of cron-utils has been
updated. This update will bring numerous amount of fixes with it
of which some will be linked in this commit message.
https://github.com/jmrozanec/cron-utils/issues/394
https://github.com/jmrozanec/cron-utils/issues/396
More to be found at: https://github.com/jmrozanec/cron-utils/compare/8.1.1...master

* Remove copyright header and jvm args (uber#417)
In the previous gradle wrapper file and gradle bat file there
were no copyright headers and no default jvm arguments.
These have been removed to comply with previous wrapper files.

* Re-added Maven publishing (uber#417)

* Renamed mavenCustom -> maven (uber#417)

* Moved maven publish to publications (uber#417)

* Removed * import from JsonDataConverter (#417)
The start import is not used anywhere else in the project, in order to
keep the code design clean, the start import has to be removed.