Activity and child workflow wait for cancellation policies (#61)
Added ability to specify behaviour of activity and child workflow in case of the enclosing CancellationScope cancellation. After this change it is possible to specify if a workflow should wait for activity confirming cancellation, cancelling it asynchronously or not cancelling it at all. For child workflows there is one more option to specify if the parent has to wait for the child confirming that it received the cancellation request.