Removes slf4j-simple from api scope (#523)
The temporal-testing module was declaring slf4j-simple in api scope, resulting
in any consumers having slf4j on their compile and runtime classpaths. Previously
there was a gradle dependency configuration rule to remove that dependency, but
I mistakenly removed that during the gradle refactoring changes in 521a992 without
noticing it in the test classpaths.

This change moves the dependency to testRuntimeOnly so that temporal-testing has
logging configured for test execution, and adds it to temporal-testing-junit4 as
testRuntimeOnly as well.

Other consumers for temporal-testing were already explicitly declaring testImplementation
dependencies on logback-classic.