Add DescribeWorkflowExecution to TestWorkflowService
This change adds support for DescribeWorkflowExecution to the
test service. It should support everything the real server
supports except for AutoResetPoints, which the test service
doesn't otherwise support (for good reason).

Implementatoin-wise, this is just a port of
HistoryEngine.DescribeWorkflowExecution from golang, except that
we infer some things from history that the real service just
stores.