Interrupt pollers when client disconnects
Connects gRPC context cancellation up to the threads performing
long-polls on task queues so that when a client exits/dies and the
attendant connections go away, the polling threads are broken out of
their blocking polls.

Also adds queue closing into the mix so that closing a workflow store
will reconfigure the queues such that they no longer block and instead
always return an empty Optional.