Introduce GH actions for native-image builds
A GH action that will build the test server native image on linux/amd64,
macOS/amd64, and windows/amd64. Build outputs are stored with the
workflow execution for 1 day and optionally attached to a user-supplied
GH release.

Missing from this work is a matrix build entry for macOS/aarch64 as
we don't have build machines for that platform yet.