Avoid spurious error notification from prep-release
The Github web ui sees the (intentional) failure of the check_release
step and generates a Big Red X warning on associated workflow runs. This
occurs even though we're indicating that an error for this step is
ok/expected via the `continue-on-error` tag. The error indication is
troubling because (a) nothing went wrong and (b) someoneone unfamiliar
with the details of the workflow might think that there's a problem.

This change uses step outputs rather than step outcomes to determine
whether or not a release already exists.