Switch from dependency version ranges to specific versions (#1314)
By using ranges we were trying to relax requirements for users who were stuck on older lib versions by providing clear boundaries of supported versions.
Dependency version ranges create too many troubles if some Temporal SDK dependency releases postfixed versions that are broken or incompatible and there is no existing solution to the problem. Which is a worse problem that the original problem we were having. It's time to move on.