Send unset catchupWindow in schedules if not specified (#2067)
If the catchupWindow property was not set in SchedulePolicy, it was sent as a zero duration in the appropriate protobuf message. However, the server expects an unset value here. The previous behavior led to setting the actual catchupWindow on the server to the minimal value (10 seconds) instead of the default (1 year).