[#1727] Use default RetryOptions for LocalActivities when not set or setMaximumAttempts(0)
Setting setMaximumAttempts=0 (or not setting any value, default = 0) for local activities should make the activity retry forever.

This is currently broken because the logic treats both no policy set and a policy with all default values (including MaxiumAttempts = 0)
as abort conditional.  What should happen (at least according the doc and to align with normal activities) is that no policy set should
be processed according to the default values of the retry policy, which includes unlimited retries when the MaxiumAttempts = 0.

Fixes #1727

Signed-off-by: Greg Haskins <greg@manetu.com>