[OpenTracing] Wrap returned promises in OutboundCallsInterceptor
This change allows the OpenTracing interceptor to transfer any active trace/span from
the point of invocation to the point of promise callback execution, thus preserving
the span within the callback.

Without this, spans become disjoint across Promise.thenCompose, due
to the nature of the deferring the execution of the .thenCompose to a callback thread.

Fixes #1962

Signed-off-by: Greg Haskins <greg@manetu.com>